Backend refactoring.
